+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 82.0.0
+## 82.0.0 (February 1, 2020)
 
 ### Enhancements
 - **Adverse Event Changes**